[hotkey] Save and present native hotkeys
Use native representation of hotkeys for both presentation and
persistence. This impacts macOS. With this approach, we can rely on
Qt to 1) give us a correct (and pretty) representation to show the user;
and 2) interpret our saved string correctly. We thus avoid a bunch of
extra conversions like cmd <-> ctrl in macOS, keeping the general
code for saving and retrieving settings clean.